requirements.txt
aiorwlock==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
appdirs-stubs==0.1.0
asn1crypto==1.5.1
async-generator==1.10
async-property==0.2.2
attrs==24.2.0
boto3==1.35.57
botocore==1.35.57
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
coverage==7.6.4
cryptography==43.0.3
cycler==0.12.1
filelock==3.16.1
firebolt-sdk==1.5.1
fonttools==4.54.1
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
readerwriterlock==1.0.9
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
sqlparse==0.5.1
tabulate==0.9.0
tomlkit==0.13.2
trio==0.27.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
